requirements.txt
openai==1.6.1
pydub
python-dotenv==1.0.0
boto3
tqdm
